requirements.txt
Flask==2.0.2
scikit-learn==1.3.2
scikit-surprise==1.1.3
pandas==2.0.3